+++ client/experimental_client_with_swig/02b_run_test_py_grabber.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-../../pm3 -c "script run testembedded_grab.py" -i