+++ client/experimental_client_with_swig/testembedded.lua
@@ -1,4 +1,13 @@
 local pm3 = require("pm3")
 p=pm3.pm3()
+
 p:console("hw status")
-print(p.name)
+p:console("hw version")
+for line in p.grabbed_output:gmatch("[^\r\n]+") do
+    if line:find("Unique ID") or line:find("uC:") then
+        print(line)
+    end
+end
+
+print("Device:", p.name)
+p:console("Rem passthru remark! :coffee:", true)